+++ nf_core/pipelines/create/create.py
@@ -101,6 +101,8 @@ class PipelineCreate:
                 ".prettierrc.yml",
                 ".github/workflows/fix-linting.yml",
             ],
+            "citations": ["assets/methods_description_template.yml"],
+            "gitpod": [".gitpod.yml"],
             "multiqc": [
                 "assets/multiqc_config.yml",
                 "assets/methods_description_template.yml",
@@ -215,6 +217,8 @@ class PipelineCreate:
             "igenomes": {"file": True, "content": True},
             "nf_core_configs": {"file": False, "content": True},
             "code_linters": {"file": True, "content": True},
+            "citations": {"file": True, "content": True},
+            "gitpod": {"file": True, "content": True},
             "multiqc": {"file": True, "content": True},
         }