+++ nf_core/pipelines/create/custompipeline.py
@@ -51,6 +51,20 @@ The available code linters are:
 - prettier (https://github.com/prettier/prettier): enforces a consistent style (indentation, quoting, line length, etc).
 """
 
+markdown_citations = """
+If adding citations, the pipeline template will contain a `CITATIONS.md` file to add the citations of all tools used in the pipeline.
+
+Additionally, it will include a YAML file (`assets/methods_description_template.yml`) to add a Materials & Methods section describing the tools used in the pieline,
+and the logics to add this section to the output MultiQC report (if the report is generated).
+"""
+
+markdown_gitpod = """
+Gitpod (https://www.gitpod.io/) provides standardized and automated development environments.
+
+Including this to your pipeline will provide an environment with the latest version of nf-core/tools installed and all its requirements.
+This is useful to have all the tools ready for pipeline development.
+"""
+
 
 class CustomPipeline(Screen):
     """Select if the pipeline will use genomic data."""
@@ -95,6 +109,18 @@ class CustomPipeline(Screen):
                 "Use code linters",
                 "The pipeline will include code linters and CI tests to lint your code: pre-commit, editor-config and prettier.",
                 "code_linters",
+            ),
+            PipelineFeature(
+                markdown_citations,
+                "Include citations",
+                "Include pipeline tools citations in CITATIONS.md and a method description in the MultiQC report (if enabled).",
+                "citations",
+            ),
+            PipelineFeature(
+                markdown_gitpod,
+                "Include a gitpod environment",
+                "Include the configuration required to use Gitpod.",
+                "gitpod",
             ),
             PipelineFeature(
                 markdown_multiqc,